+++ src/Api/Dfe.Complete.Api/Controllers/TasksDataController.cs
@@ -569,6 +569,26 @@ namespace Dfe.Complete.Api.Controllers
         {
             await sender.Send(request, cancellationToken);
             return NoContent();
+        } 
+
+        /// <summary>
+        /// Updating deed of termination for the master funding agreement for transfer project.
+        /// </summary>
+        /// <param name="request">The update command.</param>
+        /// <param name="cancellationToken">The cancellation token.</param>
+        [Authorize(Policy = "CanReadWriteUpdate")]
+        [HttpPatch]
+        [Route("TaskData/DeedOfTerminationMasterFundingAgreement")]
+        [SwaggerResponse(204, "Updated deed of termination for the master funding agreement for transfer project task successfully.")]
+        [SwaggerResponse(400, "Invalid request data.")]
+        [SwaggerResponse(404, "Project not found.")]
+        public async Task<IActionResult> UpdateDeedOfTerminationMasterFundingAgreementTaskAsync(
+            [FromBody] UpdateDeedOfTerminationMasterFundingAgreementTaskCommand request,
+            CancellationToken cancellationToken)
+        {
+            await sender.Send(request, cancellationToken);
+            return NoContent();
+        }
         }
 
         /// <summary>